refactor: extract settings feature folder
Move Settings/EditorSettings schema chunk and the settings screen
into src/app/features/settings/, keeping @/schema re-exports for
compatibility.
