+++ src/schema/index.ts
@@ -1,5 +1,9 @@
 import { Group, co, z } from "jazz-tools"
 import { Theme } from "@/app/features/themes/lib/schema"
+import {
+	Settings,
+	DEFAULT_EDITOR_SETTINGS,
+} from "@/app/features/settings/lib/schema"
 
 export {
 	ImageAsset,
@@ -10,9 +14,6 @@ export {
 	UserProfile,
 	UserRoot,
 	UserAccount,
-	Settings,
-	EditorSettings,
-	DEFAULT_EDITOR_SETTINGS,
 	migrateAnonymousData,
 	CursorEntry,
 	CursorFeed,
@@ -20,6 +21,12 @@ export {
 	createSpace,
 	createSpaceDocument,
 }
+
+export {
+	Settings,
+	EditorSettings,
+	DEFAULT_EDITOR_SETTINGS,
+} from "@/app/features/settings/lib/schema"
 
 export {
 	Theme,
@@ -34,40 +41,6 @@ let CursorEntry = z.object({
 })
 
 let CursorFeed = co.feed(CursorEntry)
-
-let StatsBadgeUnit = z.enum(["words", "sentences", "tasks"])
-
-let EditorSettings = z.object({
-	lineWidth: z.number(),
-	lineHeight: z.number(),
-	letterSpacing: z.number(),
-	fontSize: z.number(),
-	strikethroughDoneTasks: z.boolean(),
-	fadeDoneTasks: z.boolean(),
-	highlightCurrentLine: z.boolean(),
-	autoSortTasks: z.boolean(),
-	showStatsBadge: z.boolean(),
-	statsBadgeUnit: StatsBadgeUnit,
-})
-
-let DEFAULT_EDITOR_SETTINGS: z.infer<typeof EditorSettings> = {
-	lineWidth: 65,
-	lineHeight: 1.8,
-	letterSpacing: 0,
-	fontSize: 18,
-	strikethroughDoneTasks: false,
-	fadeDoneTasks: false,
-	highlightCurrentLine: true,
-	autoSortTasks: false,
-	showStatsBadge: true,
-	statsBadgeUnit: "words",
-}
-
-let Settings = co.map({
-	editor: EditorSettings,
-	defaultPreviewTheme: z.string().optional(),
-	defaultSlideshowTheme: z.string().optional(),
-})
 
 let ImageAsset = co.map({
 	type: z.literal("image"),
